hdf5-1.8.0

+++ fortran/src/H5If.c
@@ -66,7 +66,7 @@ nh5iget_name_c(hid_t_f *obj_id, _fcd buf, size_t_f *buf_size, size_t_f *name_siz
      char *c_buf =NULL;
 
      /*
-      * Allocate buffer to hold name of an object 
+      * Allocate buffer to hold name of an attribute
       */
      c_buf_size = (size_t)*buf_size;
      c_buf = (char *)HDmalloc(c_buf_size +1);
@@ -82,7 +82,7 @@ nh5iget_name_c(hid_t_f *obj_id, _fcd buf, size_t_f *buf_size, size_t_f *name_siz
      /*
       * Convert C name to FORTRAN and place it in the given buffer
       */
-      HD5packFstring(c_buf, _fcdtocp(buf), (int)c_buf_size);
+      HD5packFstring(c_buf, _fcdtocp(buf), c_buf_size);
       *name_size = (size_t_f)c_size;
       ret_value = 0;
